pyproject.toml
[tool.poetry]
name = "Diploma-project-pd12"
version = "0.1.0"
description = "todolist for SkyPro diploma project"
authors = ["Stanislav Meshcheriakov"]
readme = "README.md"
packages = [{ include = "diploma_project_pd12" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
python-dotenv = "^1.0.0"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.1"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
marshmallow = "^3.19.0"
marshmallow-dataclass = "^8.5.13"
transitions = "^0.9.0"
mypy = "^1.2.0"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.0" }
djangorestframework-stubs = { extras = ["compatible-mypy"], version = "^3.14.0" }
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest-cov = "^4.0.0"
black = "^23.3.0"
pyink = "^23.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = "True"

[tool.django-stubs]
django_settings_module = "diploma_project_pd12.settings"
